requirements.txt
baml
requests
python-dotenv
streamlit 